# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AI Content Factory 2.0 API
  version: 0.4.0
  description: P0 API contract baseline.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/projects/{projectId}/foreshadowings:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      operationId: listProjectForeshadowings
      description: Lists a project's foreshadowings, ordered by priority descending, planned_plant_chapter ascending with nulls last, then id ascending. Invalid IDs or pagination values return 400; missing projects return 404.
      parameters:
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/OffsetQuery"
      responses:
        "200": {description: Project foreshadowing list., content: {application/json: {schema: {$ref: "#/components/schemas/ForeshadowingListEnvelope"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
    post:
      operationId: createProjectForeshadowing
      description: Creates a foreshadowing. planted_plot_line_id and payoff_plot_line_id are independently optional and may each be null or reference different storylines. Each non-null reference must exist and belong to the path project; a project mismatch is 400 PLOT_LINE_PROJECT_MISMATCH. Invalid IDs or payloads return 400; missing project or referenced storyline returns 404.
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateForeshadowingRequest"}}}}
      responses:
        "201": {description: Foreshadowing created., content: {application/json: {schema: {$ref: "#/components/schemas/ForeshadowingEnvelope"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
  /api/v1/foreshadowings/{foreshadowingId}:
    parameters:
      - $ref: "#/components/parameters/ForeshadowingId"
    patch:
      operationId: updateForeshadowing
      description: Updates supplied mutable fields and status. expected_version is required; a mismatch returns 409 VERSION_CONFLICT. planted_plot_line_id and payoff_plot_line_id are independently optional; each non-null supplied reference must exist and remain in the foreshadowing project or return 400 PLOT_LINE_PROJECT_MISMATCH. Status transitions are planned to planted and planted to paid_off only; an identical update, including an unchanged status, succeeds without incrementing version. Invalid IDs or payloads return 400; missing foreshadowings or references return 404.
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/UpdateForeshadowingRequest"}}}}
      responses:
        "200": {description: Updated foreshadowing., content: {application/json: {schema: {$ref: "#/components/schemas/ForeshadowingEnvelope"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
components:
  parameters:
...
    ForeshadowingId:
      name: foreshadowingId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetQuery:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: "Invalid request. Error codes include INVALID_PROJECT_ID and VALIDATION_ERROR."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: "Resource not found. Error code: PROJECT_NOT_FOUND."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: "Optimistic concurrency conflict. Error code: VERSION_CONFLICT."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        request_id:
          type: string
    Error:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    ForeshadowingStatus:
      type: string
      enum: [planned, planted, paid_off]
    ForeshadowingPriority:
      type: string
      enum: [low, medium, high]
    Foreshadowing:
      type: object
      additionalProperties: false
      required: [id, project_id, title, description, priority, planted_plot_line_id, payoff_plot_line_id, planned_plant_chapter, planned_payoff_chapter, status, version, created_at, updated_at]
      properties:
        id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        title: {type: string, minLength: 1, maxLength: 120}
        description: {type: string, maxLength: 5000}
        priority: {$ref: "#/components/schemas/ForeshadowingPriority"}
        planted_plot_line_id: {type: [string, "null"], format: uuid, description: "Storyline where the foreshadowing is planted."}
        payoff_plot_line_id: {type: [string, "null"], format: uuid, description: "Storyline planned to pay off the foreshadowing."}
        planned_plant_chapter: {type: [integer, "null"], minimum: 1}
        planned_payoff_chapter: {type: [integer, "null"], minimum: 1}
        status: {$ref: "#/components/schemas/ForeshadowingStatus"}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
      description: planted_plot_line_id and payoff_plot_line_id are independently optional and may each be null or reference different storylines. Each non-null reference must exist and belong to project_id. When both chapter values are integers, planned_plant_chapter must not exceed planned_payoff_chapter.
    CreateForeshadowingRequest:
      type: object
      additionalProperties: false
      required: [title, description, priority, planted_plot_line_id, payoff_plot_line_id, planned_plant_chapter, planned_payoff_chapter, status]
      properties:
        title: {type: string, minLength: 1, maxLength: 120}
        description: {type: string, maxLength: 5000}
        priority: {$ref: "#/components/schemas/ForeshadowingPriority"}
        planted_plot_line_id: {type: [string, "null"], format: uuid}
        payoff_plot_line_id: {type: [string, "null"], format: uuid}
        planned_plant_chapter: {type: [integer, "null"], minimum: 1}
        planned_payoff_chapter: {type: [integer, "null"], minimum: 1}
        status: {$ref: "#/components/schemas/ForeshadowingStatus"}
      description: planted_plot_line_id and payoff_plot_line_id are independently optional. Each non-null ID must exist and belong to the path project; the two IDs may reference different storylines. When both chapter values are integers, planned_plant_chapter must not exceed planned_payoff_chapter.
    UpdateForeshadowingRequest:
      type: object
      additionalProperties: false
      required: [expected_version]
      minProperties: 2
      properties:
        expected_version: {type: integer, minimum: 1}
        title: {type: string, minLength: 1, maxLength: 120}
        description: {type: string, maxLength: 5000}
        priority: {$ref: "#/components/schemas/ForeshadowingPriority"}
        planted_plot_line_id: {type: [string, "null"], format: uuid}
        payoff_plot_line_id: {type: [string, "null"], format: uuid}
        planned_plant_chapter: {type: [integer, "null"], minimum: 1}
        planned_payoff_chapter: {type: [integer, "null"], minimum: 1}
        status: {$ref: "#/components/schemas/ForeshadowingStatus"}
      description: planted_plot_line_id and payoff_plot_line_id are independently optional. Each non-null ID must exist and remain in the foreshadowing project; the two IDs may reference different storylines. Status transitions are planned to planted and planted to paid_off only; an unchanged status is an idempotent update. When both chapter values are integers, planned_plant_chapter must not exceed planned_payoff_chapter.
    ForeshadowingList:
      type: object
      additionalProperties: false
      required: [items, total, limit, offset]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/Foreshadowing"}}
        total: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 100}
        offset: {type: integer, minimum: 0}
    ForeshadowingEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: "#/components/schemas/Foreshadowing"}
        request_id: {type: string}
    ForeshadowingListEnvelope:
      type: object
      additionalProperties: false
      required: [data, request_id]
      properties:
        data: {$ref: "#/components/schemas/ForeshadowingList"}
        request_id: {type: string}